feat: Setup project structure with clean architecture and dependency injection
- Added core modules including API, constants, validators, themes, service_locator, logger, helpers, formatters, extensions, and device_utility.
- Implemented feature modules for Google services including analytics, calendar, contacts, docs, drive, Gmail, keep, maps, meet, and sheets.
- Structured each feature module with domain, data, and presentation layers.
- Integrated get_it for dependency injection and Dio for network requests.
