+++ lib/core/utils/constants/api_constants.dart
@@ -1,0 +1,1 @@
+const String baseUrl = 'https://api.example.com';
