+++ lib/core/utils/themes/widget_themes/bottom_sheet_theme.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:google_apis_flutter/core/utils/constants/colors.dart';
+
+class AppBottomSheetTheme {
+  AppBottomSheetTheme._();
+
+  static BottomSheetThemeData lightBottomSheetTheme = BottomSheetThemeData(
+    showDragHandle: true,
+    backgroundColor: AppColors.white,
+    modalBackgroundColor: AppColors.white,
+    constraints: const BoxConstraints(minWidth: double.infinity),
+    shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+  );
+
+  static BottomSheetThemeData darkBottomSheetTheme = BottomSheetThemeData(
+    showDragHandle: true,
+    backgroundColor: AppColors.black,
+    modalBackgroundColor: AppColors.black,
+    constraints: const BoxConstraints(minWidth: double.infinity),
+    shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+  );
+}
