+++ lib/core/utils/themes/widget_themes/checkbox_theme.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+import 'package:google_apis_flutter/core/utils/constants/colors.dart';
+import 'package:google_apis_flutter/core/utils/constants/sizes.dart';
+
+/// Custom Class for Light & Dark Text Themes
+class AppCheckboxTheme {
+  AppCheckboxTheme._(); // To avoid creating instances
+
+  /// Customizable Light Text Theme
+  static CheckboxThemeData lightCheckboxTheme = CheckboxThemeData(
+    shape:
+        RoundedRectangleBorder(borderRadius: BorderRadius.circular(Sizes.xs)),
+    checkColor: WidgetStateProperty.resolveWith((states) {
+      if (states.contains(WidgetState.selected)) {
+        return AppColors.white;
+      } else {
+        return AppColors.black;
+      }
+    }),
+    fillColor: WidgetStateProperty.resolveWith((states) {
+      if (states.contains(WidgetState.selected)) {
+        return AppColors.primary;
+      } else {
+        return Colors.transparent;
+      }
+    }),
+  );
+
+  /// Customizable Dark Text Theme
+  static CheckboxThemeData darkCheckboxTheme = CheckboxThemeData(
+    shape:
+        RoundedRectangleBorder(borderRadius: BorderRadius.circular(Sizes.xs)),
+    checkColor: WidgetStateProperty.resolveWith((states) {
+      if (states.contains(WidgetState.selected)) {
+        return AppColors.white;
+      } else {
+        return AppColors.black;
+      }
+    }),
+    fillColor: WidgetStateProperty.resolveWith((states) {
+      if (states.contains(WidgetState.selected)) {
+        return AppColors.primary;
+      } else {
+        return Colors.transparent;
+      }
+    }),
+  );
+}
