+++ lib/core/utils/themes/widget_themes/chip_theme.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:google_apis_flutter/core/utils/constants/colors.dart';
+
+class AppChipTheme {
+  AppChipTheme._();
+
+  static ChipThemeData lightChipTheme = ChipThemeData(
+    disabledColor: AppColors.grey.withOpacity(0.4),
+    labelStyle: const TextStyle(color: AppColors.black),
+    selectedColor: AppColors.primary,
+    padding: const EdgeInsets.symmetric(horizontal: 12.0, vertical: 12),
+    checkmarkColor: AppColors.white,
+  );
+
+  static ChipThemeData darkChipTheme = const ChipThemeData(
+    disabledColor: AppColors.darkerGrey,
+    labelStyle: TextStyle(color: AppColors.white),
+    selectedColor: AppColors.primary,
+    padding: EdgeInsets.symmetric(horizontal: 12.0, vertical: 12),
+    checkmarkColor: AppColors.white,
+  );
+}
