+++ lib/core/utils/themes/widget_themes/outlined_button_theme.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+import 'package:google_apis_flutter/core/utils/constants/colors.dart';
+import 'package:google_apis_flutter/core/utils/constants/sizes.dart';
+
+/* -- Light & Dark Outlined Button Themes -- */
+class AppOutlinedButtonTheme {
+  AppOutlinedButtonTheme._(); //To avoid creating instances
+
+  /* -- Light Theme -- */
+  static final lightOutlinedButtonTheme = OutlinedButtonThemeData(
+    style: OutlinedButton.styleFrom(
+      elevation: 0,
+      foregroundColor: AppColors.dark,
+      side: const BorderSide(color: AppColors.borderPrimary),
+      textStyle: const TextStyle(
+          fontSize: 16, color: AppColors.black, fontWeight: FontWeight.w600),
+      padding: const EdgeInsets.symmetric(
+          vertical: Sizes.buttonHeight, horizontal: 20),
+      shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(Sizes.buttonRadius)),
+    ),
+  );
+
+  /* -- Dark Theme -- */
+  static final darkOutlinedButtonTheme = OutlinedButtonThemeData(
+    style: OutlinedButton.styleFrom(
+      foregroundColor: AppColors.light,
+      side: const BorderSide(color: AppColors.borderPrimary),
+      textStyle: const TextStyle(
+          fontSize: 16,
+          color: AppColors.textWhite,
+          fontWeight: FontWeight.w600),
+      padding: const EdgeInsets.symmetric(
+          vertical: Sizes.buttonHeight, horizontal: 20),
+      shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(Sizes.buttonRadius)),
+    ),
+  );
+}
