+++ lib/core/utils/themes/widget_themes/text_field_theme.dart
@@ -1,0 +1,74 @@
+import 'package:flutter/material.dart';
+import 'package:google_apis_flutter/core/utils/constants/colors.dart';
+import 'package:google_apis_flutter/core/utils/constants/sizes.dart';
+
+class AppTextFormFieldTheme {
+  AppTextFormFieldTheme._();
+
+  static InputDecorationTheme lightInputDecorationTheme = InputDecorationTheme(
+    errorMaxLines: 3,
+    prefixIconColor: AppColors.darkGrey,
+    suffixIconColor: AppColors.darkGrey,
+    // constraints: const BoxConstraints.expand(height: TSizes.inputFieldHeight),
+    labelStyle: const TextStyle()
+        .copyWith(fontSize: Sizes.fontSizeMd, color: AppColors.black),
+    hintStyle: const TextStyle()
+        .copyWith(fontSize: Sizes.fontSizeSm, color: AppColors.black),
+    errorStyle: const TextStyle().copyWith(fontStyle: FontStyle.normal),
+    floatingLabelStyle:
+        const TextStyle().copyWith(color: AppColors.black.withOpacity(0.8)),
+    border: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 1, color: AppColors.grey),
+    ),
+    enabledBorder: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 1, color: AppColors.grey),
+    ),
+    focusedBorder: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 1, color: AppColors.dark),
+    ),
+    errorBorder: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 1, color: AppColors.warning),
+    ),
+    focusedErrorBorder: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 2, color: AppColors.warning),
+    ),
+  );
+
+  static InputDecorationTheme darkInputDecorationTheme = InputDecorationTheme(
+    errorMaxLines: 2,
+    prefixIconColor: AppColors.darkGrey,
+    suffixIconColor: AppColors.darkGrey,
+    // constraints: const BoxConstraints.expand(height: TSizes.inputFieldHeight),
+    labelStyle: const TextStyle()
+        .copyWith(fontSize: Sizes.fontSizeMd, color: AppColors.white),
+    hintStyle: const TextStyle()
+        .copyWith(fontSize: Sizes.fontSizeSm, color: AppColors.white),
+    floatingLabelStyle:
+        const TextStyle().copyWith(color: AppColors.white.withOpacity(0.8)),
+    border: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 1, color: AppColors.darkGrey),
+    ),
+    enabledBorder: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 1, color: AppColors.darkGrey),
+    ),
+    focusedBorder: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 1, color: AppColors.white),
+    ),
+    errorBorder: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 1, color: AppColors.warning),
+    ),
+    focusedErrorBorder: const OutlineInputBorder().copyWith(
+      borderRadius: BorderRadius.circular(Sizes.inputFieldRadius),
+      borderSide: const BorderSide(width: 2, color: AppColors.warning),
+    ),
+  );
+}
