+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_apis_flutter/app.dart';
+import 'package:google_apis_flutter/core/utils/service_locator/service_locator.dart';
 
 void main() {
   runApp(const MyApp());
