--- lib/utils/helpers/helper_functions.dart
@@ -1,321 +1,0 @@
-import 'package:connectivity_plus/connectivity_plus.dart';
-import 'package:flutter/material.dart';
-import 'package:google_apis_flutter/utils/constants/colors.dart';
-import 'package:google_apis_flutter/utils/constants/enums.dart';
-import 'package:intl/intl.dart';
-
-class HelperFunctions {
-  static void showSnackBar({
-    required String message,
-    required BuildContext context,
-    SnackBarType type = SnackBarType.error,
-    Duration duration = const Duration(seconds: 5),
-    double elevation = 6.0,
-    EdgeInsetsGeometry margin = const EdgeInsets.all(10.0),
-    double borderRadius = 8.0,
-    TextStyle textStyle = const TextStyle(color: Colors.white),
-    Color actionTextColor = Colors.white,
-    SnackBarAction? action,
-  }) {
-    // Define custom icons for different SnackBar types
-    IconData iconData;
-    Color backgroundColor;
-    Color iconColor;
-
-    switch (type) {
-      case SnackBarType.error:
-        iconData = Icons.error;
-        backgroundColor = Colors.red;
-        iconColor = Colors.white;
-        break;
-      case SnackBarType.success:
-        iconData = Icons.check_circle;
-        backgroundColor = Colors.green;
-        iconColor = Colors.white;
-        break;
-      case SnackBarType.info:
-        iconData = Icons.info;
-        backgroundColor = Colors.blue;
-        iconColor = Colors.white;
-        break;
-    }
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Row(
-          children: [
-            Icon(
-              iconData,
-              color: iconColor,
-            ),
-            const SizedBox(width: 10),
-            Expanded(
-              child: Text(
-                message,
-                style: textStyle,
-              ),
-            ),
-          ],
-        ),
-        duration: duration,
-        backgroundColor: backgroundColor,
-        action: action,
-        elevation: elevation,
-        margin: margin,
-        behavior: SnackBarBehavior.floating,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(borderRadius),
-        ),
-      ),
-    );
-  }
-
-  static void showAlert({
-    required String title,
-    required String message,
-    required BuildContext context,
-    AlertType type = AlertType.info,
-  }) {
-    Color iconColor;
-    switch (type) {
-      case AlertType.success:
-        iconColor = AppColors.success;
-        break;
-      case AlertType.error:
-        iconColor = AppColors.error;
-        break;
-      default:
-        iconColor = AppColors.info;
-    }
-
-    showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return Dialog(
-          backgroundColor: Colors.transparent,
-          insetPadding: const EdgeInsets.symmetric(horizontal: 20),
-          child: Stack(
-            alignment: Alignment.center,
-            children: [
-              Container(
-                padding: const EdgeInsets.all(20),
-                margin: const EdgeInsets.only(top: 20),
-                decoration: BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.circular(12),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.black.withOpacity(0.1),
-                      spreadRadius: 1,
-                      blurRadius: 10,
-                      offset: const Offset(0, 5),
-                    ),
-                  ],
-                ),
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    TweenAnimationBuilder(
-                      tween: Tween<double>(
-                          begin: 0.5, end: 1.0), // Change the scale factor
-                      duration: const Duration(
-                          milliseconds: 500), // Increase the duration
-                      curve: Curves
-                          .easeInOut, // Use elasticOut curve for a bouncy effect
-                      builder: (context, value, child) {
-                        return Transform.scale(
-                          scale: value,
-                          child: child,
-                        );
-                      },
-                      child: Icon(
-                        type == AlertType.success
-                            ? Icons.check_circle
-                            : type == AlertType.error
-                                ? Icons.error_outline
-                                : Icons.info_outline,
-                        color: iconColor,
-                        size: 64,
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    Text(
-                      title,
-                      style: const TextStyle(
-                        fontSize: 20,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                    const SizedBox(height: 8),
-                    Text(
-                      message,
-                      textAlign: TextAlign.center,
-                      style: const TextStyle(
-                        fontSize: 16,
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    SizedBox(
-                      width: double.infinity,
-                      child: ElevatedButton(
-                        onPressed: () => Navigator.of(context).pop(),
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: iconColor,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(20),
-                          ),
-                          padding: const EdgeInsets.symmetric(vertical: 12),
-                        ),
-                        child: const Text(
-                          'OK',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 16,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-
-  static void navigateToScreen(BuildContext context, Widget screen) {
-    Navigator.push(
-      context,
-      PageRouteBuilder(
-        pageBuilder: (_, __, ___) => screen,
-        transitionsBuilder: (_, animation, __, child) {
-          return ScaleTransition(
-            scale: Tween<double>(begin: 0.5, end: 1.0).animate(
-              CurvedAnimation(
-                parent: animation,
-                curve: Curves.fastOutSlowIn,
-              ),
-            ),
-            child: SlideTransition(
-              position: Tween<Offset>(
-                begin: const Offset(0.0, 1.0),
-                end: Offset.zero,
-              ).animate(
-                CurvedAnimation(
-                  parent: animation,
-                  curve: Curves.fastOutSlowIn,
-                ),
-              ),
-              child: FadeTransition(
-                opacity: animation,
-                child: child,
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  static void popScreen(BuildContext context) {
-    Navigator.pop(context);
-    _playfulPopAnimation(context);
-  }
-
-  static void _playfulPopAnimation(BuildContext context) {
-    Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(
-        parent: ModalRoute.of(context)!.animation!,
-        curve: Curves.elasticOut, // Add a playful bounce effect
-        reverseCurve: Curves.easeOutBack, // Smooth reverse animation
-      ),
-    );
-  }
-
-  static void navigateReplacementToScreen(BuildContext context, Widget screen) {
-    Navigator.pushReplacement(
-      context,
-      PageRouteBuilder(
-        pageBuilder: (_, __, ___) => screen,
-        transitionsBuilder: (_, animation, __, child) {
-          return ScaleTransition(
-            scale: Tween<double>(begin: 0.5, end: 1.0).animate(
-              CurvedAnimation(
-                parent: animation,
-                curve: Curves.fastOutSlowIn,
-              ),
-            ),
-            child: SlideTransition(
-              position: Tween<Offset>(
-                begin: const Offset(0.0, 1.0),
-                end: Offset.zero,
-              ).animate(
-                CurvedAnimation(
-                  parent: animation,
-                  curve: Curves.fastOutSlowIn,
-                ),
-              ),
-              child: FadeTransition(
-                opacity: animation,
-                child: child,
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  static String truncateText(String text, int maxLength) {
-    if (text.length <= maxLength) {
-      return text;
-    } else {
-      return '${text.substring(0, maxLength)}...';
-    }
-  }
-
-  static bool isDarkMode(BuildContext context) {
-    return Theme.of(context).brightness == Brightness.dark;
-  }
-
-  static Size screenSize(BuildContext context) {
-    return MediaQuery.of(context).size;
-  }
-
-  static double screenHeight(BuildContext context) {
-    return MediaQuery.sizeOf(context).height;
-  }
-
-  static double screenWidth(BuildContext context) {
-    return MediaQuery.sizeOf(context).width;
-  }
-
-  static String getFormattedDate(DateTime date,
-      {String format = 'dd MMM yyyy'}) {
-    return DateFormat(format).format(date);
-  }
-
-  static List<T> removeDuplicates<T>(List<T> list) {
-    return list.toSet().toList();
-  }
-
-  static List<Widget> wrapWidgets(List<Widget> widgets, int rowSize) {
-    final wrappedList = <Widget>[];
-    for (var i = 0; i < widgets.length; i += rowSize) {
-      final rowChildren = widgets.sublist(
-          i, i + rowSize > widgets.length ? widgets.length : i + rowSize);
-      wrappedList.add(Row(children: rowChildren));
-    }
-    return wrappedList;
-  }
-
-  static final Connectivity _connectivity = Connectivity();
-
-  static Future<bool> isConnected() async {
-    var connectivityResult = await _connectivity.checkConnectivity();
-    bool isConnected = connectivityResult != ConnectivityResult.none;
-    return isConnected;
-  }
-}
