--- lib/utils/themes/widget_themes/elevated_button_theme.dart
@@ -1,46 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:google_apis_flutter/utils/constants/colors.dart';
-import 'package:google_apis_flutter/utils/constants/sizes.dart';
-
-/* -- Light & Dark Elevated Button Themes -- */
-class AppElevatedButtonTheme {
-  AppElevatedButtonTheme._(); //To avoid creating instances
-
-  /* -- Light Theme -- */
-  static final lightElevatedButtonTheme = ElevatedButtonThemeData(
-    style: ElevatedButton.styleFrom(
-      elevation: 0,
-      foregroundColor: AppColors.light,
-      backgroundColor: AppColors.primary,
-      disabledForegroundColor: AppColors.darkGrey,
-      disabledBackgroundColor: AppColors.buttonDisabled,
-      side: const BorderSide(color: AppColors.primary),
-      padding: const EdgeInsets.symmetric(vertical: Sizes.buttonHeight),
-      textStyle: const TextStyle(
-          fontSize: 16,
-          color: AppColors.textWhite,
-          fontWeight: FontWeight.w600),
-      shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(Sizes.buttonRadius)),
-    ),
-  );
-
-  /* -- Dark Theme -- */
-  static final darkElevatedButtonTheme = ElevatedButtonThemeData(
-    style: ElevatedButton.styleFrom(
-      elevation: 0,
-      foregroundColor: AppColors.light,
-      backgroundColor: AppColors.primary,
-      disabledForegroundColor: AppColors.darkGrey,
-      disabledBackgroundColor: AppColors.darkerGrey,
-      side: const BorderSide(color: AppColors.primary),
-      padding: const EdgeInsets.symmetric(vertical: Sizes.buttonHeight),
-      textStyle: const TextStyle(
-          fontSize: 16,
-          color: AppColors.textWhite,
-          fontWeight: FontWeight.w600),
-      shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(Sizes.buttonRadius)),
-    ),
-  );
-}
